requirements.txt
cryptography==2.8
Django==3.0.4
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-dotenv==0.10.1
social-auth-app-django==3.1.0
djangorestframework